# Conflicts:
#	.idea/workspace.xml
#	athlete/athlete_complaint.cpp
#	cmake-build-debug/2024_NJUPT_Program_Design.exe
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/3H5RFNM4OQ/moc_athlete_check_results.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/3H5RFNM4OQ/moc_athlete_complaint.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/3H5RFNM4OQ/moc_athlete_mainwindow.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_management_of_competition.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_management_of_competition_cell.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_management_of_competiton_search.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_management_of_participating_departments.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_sys_admin_check_results.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_sys_admin_handle_complaint.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_sys_admin_mainwindow.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/BT4GILMY53/moc_sys_admin_review_results.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/CWNBZNAZQH/moc_referee_handle_complaint.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/CWNBZNAZQH/moc_referee_handle_complaint_cell.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/CWNBZNAZQH/moc_referee_mainwindow.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/CWNBZNAZQH/moc_referee_submit_results.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/CWNBZNAZQH/moc_referee_submit_results_cell.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/EWIEGA46WW/moc_login.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/IVBR36VHBM/moc_college_admin_check_results.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/IVBR36VHBM/moc_college_admin_complaint.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/IVBR36VHBM/moc_college_mainwindow.cpp.d
#	cmake-build-debug/2024_NJUPT_Program_Design_autogen/deps
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/athlete/athlete_complaint.cpp.obj
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/objects.a
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/referee/referee_handle_complaint.cpp.obj
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/referee/referee_submit_results.cpp.obj
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/referee/referee_submit_results.cpp.obj.d
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/sys_admin/management_of_competition.cpp.obj
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design.dir/sys_admin/management_of_competition.cpp.obj.d
#	cmake-build-debug/CMakeFiles/2024_NJUPT_Program_Design_autogen.dir/AutogenInfo.json
#	cmake-build-debug/CMakeFiles/3.29.3/CMakeDetermineCompilerABI_C.bin
#	cmake-build-debug/CMakeFiles/3.29.3/CMakeDetermineCompilerABI_CXX.bin
#	cmake-build-debug/CMakeFiles/3.29.3/CompilerIdC/a.exe
#	cmake-build-debug/CMakeFiles/3.29.3/CompilerIdCXX/a.exe
#	cmake-build-debug/CMakeFiles/CMakeConfigureLog.yaml
#	cmake-build-debug/CMakeFiles/Makefile.cmake
#	cmake-build-debug/CMakeFiles/clion-Debug-log.txt
#	cmake-build-debug/Testing/Temporary/LastTest.log
#	sys_admin/management_of_competition.cpp

+++ referee/referee_submit_results.cpp
@@ -9,7 +9,7 @@
 
 #include "referee_submit_results.h"
 #include "ui_referee_submit_results.h"
-
+#include <QMessageBox>
 
 referee_submit_results::referee_submit_results(QWidget *parent) :
     QWidget(parent), ui(new Ui::referee_submit_results) {
@@ -35,6 +35,11 @@ void referee_submit_results::slot1() {
 
 void referee_submit_results::slot2() {
     int curRow = ui->tableWidget->currentRow();
+
+    if (curRow < 0) {
+        QMessageBox::warning(this, tr("删除失败"), tr("请选择一行进行删除！"), QMessageBox::Ok);
+        return;
+    }
     queryString = QString("delete from results where 运动员号码 = '%1' and 项目编号 = '%2'")
             .arg(ui->tableWidget->item(curRow, 0)->text())
             .arg(ui->tableWidget->item(curRow, 3)->text());
